Login Page V0.1 + Main Application V0.1
I made a login page which lets you log in by checking whether or you are entering the right master password or not(checks file where master password is stored). Once you have logged in successfully, you go to the main application page, the main application page is missing a lot of stuff right now, since I just made a rough version of it. Currently the only thing that half works is the add login function, with that you can add your login which gets stored into a file(this file will later have the login data encrypted). This function does store the username and password successfully but for some reason it doesn't store the site name. Will try to fix this by the next version.