Login Page V0.3 + Main Application V0.4
Today I fixed some positioning issues I was facing whenever the add login fields frame was called, to be specific, when that frame was called the application’s logo and name labels were being pushed aside to the right as well as the password generator’s GUI frame. To fix this issue I used .place instead of .grid which allowed me to manually change every widgets position freely and precisely. In addition to this, I modified with the configuration of the frames for the widgets for aesthetic purposes and alongside this I also redid the login page for better aesthetics, to do this I added some new labels and changed the fonts of the GUI in the page and made the frame spawn in the middle of the page.