View Accounts V0.1
Working on the View Accounts part of the application. Currently its not working as it is still being made, I added it here just to save the progress and have a checkpoint to go back to. The only thing I changed that is working is the part where now all the login information is saved in their own separate files instead of just all in one file like previously.