View Accounts V.03 + minor changes
The view accounts part is now functional, the only thing that needs to be finished for that bit is the alignment of the frame for view accounts and the frame in which you can see the account information. Minor changes include me adding the applications name in the sign up page for purely aesthetic reasons, and adding a word limit for every entry field except for passwords in the sign up page(master password).